% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airLDN.R
\docType{data}
\name{airLDN}
\alias{airLDN}
\title{Daily Observations of the Air Quality Index of London (Britain)}
\format{
A data frame with 2552 rows and 8 variables:
\describe{
  \item{date}{the observation date}
  \item{pm25}{particle matter with an aerodynamic diameter smaller than 2.5
  \eqn{\mu}m}
  \item{pm10}{particle matter with an aerodynamic diameter smaller than 10
  \eqn{\mu}m}
  \item{o3}{ozone or trioxygen}
  \item{no2}{nitrogen dioxide}
  \item{so2}{sulphur dioxide}
  \item{co}{carbon monoxide}
  \item{AQI}{composite air quality index}
}
}
\source{
The data can be obtained from the World Air Quality Project.

\url{https://aqicn.org/city/london}
}
\usage{
airLDN
}
\description{
A dataset that contains daily observations of individual air pollutants
from 2014 to 2020.
}
\keyword{datasets}
