% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpG7.R
\docType{data}
\name{gdpG7}
\alias{gdpG7}
\title{Quarterly G7 GDP, Q1 1962 to Q4 2019}
\format{
A data frame with 232 rows and 3 variables:
\describe{
  \item{year}{the observation year}
  \item{quarter}{the observation quarter in the given year}
  \item{gdp}{the volume Index of the Gross Domestic Product of the G7 nations}
}
}
\source{
The data was obtained from the Organization for Economic Co-operation
and Development (OECD)

\url{https://data.oecd.org/gdp/quarterly-gdp.htm#indicator-chart}
}
\usage{
gdpG7
}
\description{
A dataset that contains the (seasonally adjusted) Gross
Domestic Product of the G7 nations from the first quarter of 1962 to the
fourth quarter of 2019
}
\keyword{datasets}
