% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.esemifar.R
\name{plot.esemifar}
\alias{plot.esemifar}
\title{Plot Method for the Package 'esemifar'}
\usage{
\method{plot}{esemifar}(x, t = NULL, ...)
}
\arguments{
\item{x}{an input object of class \code{esemifar}.}

\item{t}{an optional vector with time points that will be considered for
the x-axis within the plot; is set to NULL by default and uses a vector
\code{1:length(x$ye)} for time points.}

\item{...}{additional arguments of the standard plot method.}
}
\value{
None
}
\description{
This function regulates how objects created by the package \code{esemifar} are
plotted.
}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
}
}
