% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.esemifar.R
\name{print.esemifar}
\alias{print.esemifar}
\title{Print Method for the Package 'esemifar'}
\usage{
\method{print}{esemifar}(x, ...)
}
\arguments{
\item{x}{an input object of class \code{esemifar}.}

\item{...}{included for compatibility; additional arguments will however
not affect the output.}
}
\value{
None
}
\description{
This function regulates how objects created by the package \code{esemifar} are
printed.
}
\author{
\itemize{
\item Dominik Schulz (Scientific employee) (Department of Economics, Paderborn
University), \cr
}
}
